+++ src/pages/AttendancePage/index.js
@@ -1,0 +1,2 @@
+export { default as AttendanceSignPage } from './AttendanceSignPage';
+export { default as AttendanceStudentIdPage } from './AttendanceStudentIdPage';
