+++ src/pages/DashboardPage/index.js
@@ -1,4 +1,4 @@
 export { default as DashboardPage } from './DashboardPage';
-export { default as DashboardInfo } from './DashboardInfo';
-export { default as DashboardEmail } from './DashboardEmail';
+export { default as DashboardInfoPage } from './DashboardInfoPage';
+export { default as DashboardEmailPage } from './DashboardEmailPage';
 export { default as DashboardAttendeePage } from './DashboardAttendeePage';
